backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.2
anthropic==0.34.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
playwright==1.58.0
apscheduler==3.10.4
